refactor, tested view
